requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.7.2
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.1.11
Django==3.2.6
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
django-taggit==1.5.1
django-taggit-serializer==0.1.7
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pep8==1.7.1
Pillow==8.3.2
platformdirs==2.2.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.10.2
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
